Revert "Hybrid Consensus Node"
