+++ node/src/cli.rs
@@ -1,14 +1,5 @@
-use crate::{
-    client::{FullBackend, FullClient},
-    consensus::{AuraConsensus, BabeConsensus},
-    ethereum::{EthConfiguration, FrontierBackend},
-    service::new_chain_ops,
-};
-use node_subtensor_runtime::opaque::Block;
+use crate::ethereum::EthConfiguration;
 use sc_cli::RunCmd;
-use sc_consensus::BasicQueue;
-use sc_service::{Configuration, TaskManager};
-use std::sync::Arc;
 
 #[derive(Debug, clap::Parser)]
 pub struct Cli {
@@ -21,15 +12,6 @@ pub struct Cli {
     /// Choose sealing method.
     #[arg(long, value_enum, ignore_case = true)]
     pub sealing: Option<Sealing>,
-
-    /// Whether to try Aura or Babe consensus on first start.
-    ///
-    /// After starting, the consensus used by the node will automatically
-    /// switch to whatever is required to continue validating / syncing.
-    ///
-    /// TODO: Remove this after the Babe transition has settled.
-    #[arg(long, value_enum, ignore_case = true, default_value_t=SupportedConsensusMechanism::default())]
-    pub initial_consensus: SupportedConsensusMechanism,
 
     #[command(flatten)]
     pub eth: EthConfiguration,
@@ -80,38 +62,4 @@ pub enum Sealing {
     Manual,
     /// Seal when transaction is executed.
     Instant,
-}
-
-/// Supported consensus mechanisms.
-#[derive(Copy, Clone, Debug, Default, clap::ValueEnum)]
-pub enum SupportedConsensusMechanism {
-    // Babe
-    Babe,
-    /// Aura
-    #[default]
-    Aura,
-}
-
-// Convinience methods for static dispatch of different service methods with
-// different consensus mechanisms.
-impl SupportedConsensusMechanism {
-    pub fn new_chain_ops(
-        &self,
-        config: &mut Configuration,
-        eth_config: &EthConfiguration,
-    ) -> Result<
-        (
-            Arc<FullClient>,
-            Arc<FullBackend>,
-            BasicQueue<Block>,
-            TaskManager,
-            FrontierBackend,
-        ),
-        sc_service::Error,
-    > {
-        match self {
-            SupportedConsensusMechanism::Aura => new_chain_ops::<AuraConsensus>(config, eth_config),
-            SupportedConsensusMechanism::Babe => new_chain_ops::<BabeConsensus>(config, eth_config),
-        }
-    }
 }
