--- node/src/consensus/mod.rs
@@ -1,9 +1,0 @@
-mod aura_consensus;
-mod aura_wrapped_import_queue;
-mod babe_consensus;
-mod consensus_mechanism;
-
-pub use aura_consensus::AuraConsensus;
-pub use babe_consensus::BabeConsensus;
-pub use consensus_mechanism::ConsensusMechanism;
-pub use consensus_mechanism::StartAuthoringParams;
