+++ node/src/ethereum.rs
@@ -1,4 +1,5 @@
-use crate::rpc::EthDeps;
+use fc_aura::AuraConsensusDataProvider;
+pub use fc_consensus::FrontierBlockImport;
 use fc_rpc::{
     Debug, DebugApiServer, Eth, EthApiServer, EthConfig, EthDevSigner, EthFilter,
     EthFilterApiServer, EthPubSub, EthPubSubApiServer, EthSigner, EthTask, Net, NetApiServer, Web3,
@@ -27,6 +28,7 @@ use std::{
 };
 
 use crate::client::{FullBackend, FullClient};
+use crate::rpc::EthDeps;
 
 pub type FrontierBackend = fc_db::Backend<Block, FullClient>;
 
@@ -196,7 +198,6 @@ pub async fn spawn_frontier_tasks(
 fn extend_rpc_aet_api<P, CT, CIDP, EC>(
     io: &mut RpcModule<()>,
     deps: &EthDeps<P, CT, CIDP>,
-    pending_consensus_data_provider: Option<Box<dyn fc_rpc::pending::ConsensusDataProvider<Block>>>,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>>
 where
     P: TransactionPool<
@@ -232,7 +233,9 @@ where
             deps.execute_gas_limit_multiplier,
             deps.forced_parent_hashes.clone(),
             deps.pending_create_inherent_data_providers.clone(),
-            pending_consensus_data_provider,
+            Some(Box::new(AuraConsensusDataProvider::new(
+                deps.client.clone(),
+            ))),
         )
         .replace_config::<EC>()
         .into_rpc(),
@@ -390,7 +393,6 @@ pub fn create_eth<P, CT, CIDP, EC>(
             fc_mapping_sync::EthereumBlockNotification<Block>,
         >,
     >,
-    pending_consensus_data_provider: Option<Box<dyn fc_rpc::pending::ConsensusDataProvider<Block>>>,
 ) -> Result<RpcModule<()>, Box<dyn std::error::Error + Send + Sync>>
 where
     P: TransactionPool<
@@ -404,7 +406,7 @@ where
     CIDP: CreateInherentDataProviders<Block, ()> + Send + Clone + 'static,
     EC: EthConfig<Block, FullClient>,
 {
-    extend_rpc_aet_api::<P, CT, CIDP, EC>(&mut io, &deps, pending_consensus_data_provider)?;
+    extend_rpc_aet_api::<P, CT, CIDP, EC>(&mut io, &deps)?;
     extend_rpc_eth_filter::<P, CT, CIDP>(&mut io, &deps)?;
     extend_rpc_eth_pubsub::<P, CT, CIDP>(
         &mut io,
