+++ node/src/lib.rs
@@ -1,8 +1,6 @@
 pub mod chain_spec;
 pub mod cli;
 pub mod client;
-pub mod conditional_evm_block_import;
-pub mod consensus;
 pub mod ethereum;
 pub mod rpc;
 pub mod service;
