+++ node/src/service.rs
@@ -1,60 +1,48 @@
 //! Service and ServiceFactory implementation. Specialized wrapper over substrate service.
 
-use crate::consensus::ConsensusMechanism;
 use futures::{FutureExt, channel::mpsc, future};
 use node_subtensor_runtime::{RuntimeApi, TransactionConverter, opaque::Block};
 use sc_chain_spec::ChainType;
 use sc_client_api::{Backend as BackendT, BlockBackend};
-use sc_consensus::{BasicQueue, BoxBlockImport};
+use sc_consensus::{
+    BasicQueue, BlockCheckParams, BlockImport, BlockImportParams, BoxBlockImport, ImportResult,
+};
 use sc_consensus_grandpa::BlockNumberOps;
 use sc_consensus_slots::BackoffAuthoringOnFinalizedHeadLagging;
-use sc_consensus_slots::SlotProportion;
 use sc_network::config::SyncMode;
 use sc_network_sync::strategy::warp::{WarpSyncConfig, WarpSyncProvider};
 use sc_service::{Configuration, PartialComponents, TaskManager, error::Error as ServiceError};
 use sc_telemetry::{Telemetry, TelemetryHandle, TelemetryWorker, log};
 use sc_transaction_pool::TransactionPoolHandle;
 use sc_transaction_pool_api::OffchainTransactionPoolFactory;
+use sp_consensus::Error as ConsensusError;
+use sp_consensus_aura::sr25519::AuthorityPair as AuraPair;
 use sp_core::H256;
-use sp_runtime::traits::{Block as BlockT, NumberFor};
+use sp_runtime::traits::{Block as BlockT, Header, NumberFor};
 use std::collections::HashSet;
 use std::str::FromStr;
-use std::sync::atomic::AtomicBool;
 use std::{cell::RefCell, path::Path};
-use std::{sync::Arc, time::Duration};
+use std::{marker::PhantomData, sync::Arc, time::Duration};
 use substrate_prometheus_endpoint::Registry;
 
 use crate::cli::Sealing;
 use crate::client::{FullBackend, FullClient, HostFunctions, RuntimeExecutor};
 use crate::ethereum::{
-    BackendType, EthConfiguration, FrontierBackend, FrontierPartialComponents, StorageOverride,
-    StorageOverrideHandler, db_config_dir, new_frontier_partial, spawn_frontier_tasks,
+    BackendType, EthConfiguration, FrontierBackend, FrontierBlockImport, FrontierPartialComponents,
+    StorageOverride, StorageOverrideHandler, db_config_dir, new_frontier_partial,
+    spawn_frontier_tasks,
 };
 
 /// The minimum period of blocks on which justifications will be
 /// imported and generated.
 const GRANDPA_JUSTIFICATION_PERIOD: u32 = 512;
 
-pub type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
-pub type GrandpaBlockImport =
+type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
+type GrandpaBlockImport =
     sc_consensus_grandpa::GrandpaBlockImport<FullBackend, Block, FullClient, FullSelectChain>;
 type GrandpaLinkHalf = sc_consensus_grandpa::LinkHalf<Block, FullClient, FullSelectChain>;
-#[allow(clippy::upper_case_acronyms)]
-pub type BIQ<'a> = Box<
-    dyn FnOnce(
-            Arc<FullClient>,
-            Arc<FullBackend>,
-            &Configuration,
-            &EthConfiguration,
-            &TaskManager,
-            Option<TelemetryHandle>,
-            GrandpaBlockImport,
-            Arc<TransactionPoolHandle<Block, FullClient>>,
-        ) -> Result<(BasicQueue<Block>, BoxBlockImport<Block>), sc_service::Error>
-        + 'a,
->;
 
-pub fn new_partial(
+pub fn new_partial<BIQ>(
     config: &Configuration,
     eth_config: &EthConfiguration,
     build_import_queue: BIQ,
@@ -74,7 +62,17 @@ pub fn new_partial(
         ),
     >,
     ServiceError,
-> {
+>
+where
+    BIQ: FnOnce(
+        Arc<FullClient>,
+        &Configuration,
+        &EthConfiguration,
+        &TaskManager,
+        Option<TelemetryHandle>,
+        GrandpaBlockImport,
+    ) -> Result<(BasicQueue<Block>, BoxBlockImport<Block>), ServiceError>,
+{
     let telemetry = config
         .telemetry_endpoints
         .clone()
@@ -168,6 +166,15 @@ pub fn new_partial(
         }
     };
 
+    let (import_queue, block_import) = build_import_queue(
+        client.clone(),
+        config,
+        eth_config,
+        &task_manager,
+        telemetry.as_ref().map(|x| x.handle()),
+        grandpa_block_import,
+    )?;
+
     let transaction_pool = Arc::from(
         sc_transaction_pool::Builder::new(
             task_manager.spawn_essential_handle(),
@@ -178,17 +185,6 @@ pub fn new_partial(
         .with_prometheus(config.prometheus_registry())
         .build(),
     );
-
-    let (import_queue, block_import) = build_import_queue(
-        client.clone(),
-        backend.clone(),
-        config,
-        eth_config,
-        &task_manager,
-        telemetry.as_ref().map(|x| x.handle()),
-        grandpa_block_import,
-        transaction_pool.clone(),
-    )?;
 
     Ok(PartialComponents {
         client,
@@ -208,24 +204,131 @@ pub fn new_partial(
     })
 }
 
+pub struct ConditionalEVMBlockImport<B: BlockT, I, F> {
+    inner: I,
+    frontier_block_import: F,
+    _marker: PhantomData<B>,
+}
+
+impl<B, I, F> Clone for ConditionalEVMBlockImport<B, I, F>
+where
+    B: BlockT,
+    I: Clone + BlockImport<B>,
+    F: Clone + BlockImport<B>,
+{
+    fn clone(&self) -> Self {
+        ConditionalEVMBlockImport {
+            inner: self.inner.clone(),
+            frontier_block_import: self.frontier_block_import.clone(),
+            _marker: PhantomData,
+        }
+    }
+}
+
+impl<B, I, F> ConditionalEVMBlockImport<B, I, F>
+where
+    B: BlockT,
+    I: BlockImport<B>,
+    I::Error: Into<ConsensusError>,
+    F: BlockImport<B>,
+    F::Error: Into<ConsensusError>,
+{
+    pub fn new(inner: I, frontier_block_import: F) -> Self {
+        Self {
+            inner,
+            frontier_block_import,
+            _marker: PhantomData,
+        }
+    }
+}
+
+#[async_trait::async_trait]
+impl<B, I, F> BlockImport<B> for ConditionalEVMBlockImport<B, I, F>
+where
+    B: BlockT,
+    I: BlockImport<B> + Send + Sync,
+    I::Error: Into<ConsensusError>,
+    F: BlockImport<B> + Send + Sync,
+    F::Error: Into<ConsensusError>,
+{
+    type Error = ConsensusError;
+
+    async fn check_block(&self, block: BlockCheckParams<B>) -> Result<ImportResult, Self::Error> {
+        self.inner.check_block(block).await.map_err(Into::into)
+    }
+
+    async fn import_block(&self, block: BlockImportParams<B>) -> Result<ImportResult, Self::Error> {
+        // 4345556 - mainnet runtime upgrade block with Frontier
+        if *block.header.number() < 4345557u32.into() {
+            self.inner.import_block(block).await.map_err(Into::into)
+        } else {
+            self.frontier_block_import
+                .import_block(block)
+                .await
+                .map_err(Into::into)
+        }
+    }
+}
+
+/// Build the import queue for the template runtime (aura + grandpa).
+pub fn build_aura_grandpa_import_queue(
+    client: Arc<FullClient>,
+    config: &Configuration,
+    _eth_config: &EthConfiguration,
+    task_manager: &TaskManager,
+    telemetry: Option<TelemetryHandle>,
+    grandpa_block_import: GrandpaBlockImport,
+) -> Result<(BasicQueue<Block>, BoxBlockImport<Block>), ServiceError>
+where
+    NumberFor<Block>: BlockNumberOps,
+{
+    let conditional_block_import = ConditionalEVMBlockImport::new(
+        grandpa_block_import.clone(),
+        FrontierBlockImport::new(grandpa_block_import.clone(), client.clone()),
+    );
+
+    let slot_duration = sc_consensus_aura::slot_duration(&*client)?;
+    let create_inherent_data_providers = move |_, ()| async move {
+        let timestamp = sp_timestamp::InherentDataProvider::from_system_time();
+        let slot =
+            sp_consensus_aura::inherents::InherentDataProvider::from_timestamp_and_slot_duration(
+                *timestamp,
+                slot_duration,
+            );
+        Ok((slot, timestamp))
+    };
+
+    let import_queue = sc_consensus_aura::import_queue::<AuraPair, _, _, _, _, _>(
+        sc_consensus_aura::ImportQueueParams {
+            block_import: conditional_block_import.clone(),
+            justification_import: Some(Box::new(grandpa_block_import.clone())),
+            client,
+            create_inherent_data_providers,
+            spawner: &task_manager.spawn_essential_handle(),
+            registry: config.prometheus_registry(),
+            check_for_equivocation: Default::default(),
+            telemetry,
+            compatibility_mode: sc_consensus_aura::CompatibilityMode::None,
+        },
+    )
+    .map_err::<ServiceError, _>(Into::into)?;
+
+    Ok((import_queue, Box::new(conditional_block_import)))
+}
+
 /// Build the import queue for the template runtime (manual seal).
-#[allow(clippy::too_many_arguments)]
-#[cfg(feature = "runtime-benchmarks")]
 pub fn build_manual_seal_import_queue(
     client: Arc<FullClient>,
-    _backend: Arc<FullBackend>,
     config: &Configuration,
     _eth_config: &EthConfiguration,
     task_manager: &TaskManager,
     _telemetry: Option<TelemetryHandle>,
     grandpa_block_import: GrandpaBlockImport,
-    _transaction_pool_handle: Arc<TransactionPoolHandle<Block, FullClient>>,
 ) -> Result<(BasicQueue<Block>, BoxBlockImport<Block>), ServiceError> {
-    let conditional_block_import =
-        crate::conditional_evm_block_import::ConditionalEVMBlockImport::new(
-            grandpa_block_import.clone(),
-            fc_consensus::FrontierBlockImport::new(grandpa_block_import.clone(), client.clone()),
-        );
+    let conditional_block_import = ConditionalEVMBlockImport::new(
+        grandpa_block_import.clone(),
+        FrontierBlockImport::new(grandpa_block_import.clone(), client.clone()),
+    );
     Ok((
         sc_consensus_manual_seal::import_queue(
             Box::new(conditional_block_import.clone()),
@@ -237,16 +340,14 @@ pub fn build_manual_seal_import_queue(
 }
 
 /// Builds a new service for a full client.
-pub async fn new_full<NB, CM>(
+pub async fn new_full<NB>(
     mut config: Configuration,
     eth_config: EthConfiguration,
     sealing: Option<Sealing>,
-    custom_service_signal: Option<Arc<AtomicBool>>,
 ) -> Result<TaskManager, ServiceError>
 where
     NumberFor<Block>: BlockNumberOps,
     NB: sc_network::NetworkBackend<Block, <Block as BlockT>::Hash>,
-    CM: ConsensusMechanism,
 {
     // Substrate doesn't seem to support fast sync option in our configuration.
     if matches!(config.network.sync_mode, SyncMode::LightState { .. }) {
@@ -257,8 +358,11 @@ where
         return Err(ServiceError::Other("Unsupported sync mode".to_string()));
     }
 
-    let mut consensus_mechanism = CM::new();
-    let build_import_queue = consensus_mechanism.build_biq()?;
+    let build_import_queue = if sealing.is_some() {
+        build_manual_seal_import_queue
+    } else {
+        build_aura_grandpa_import_queue
+    };
 
     let PartialComponents {
         client,
@@ -270,12 +374,6 @@ where
         transaction_pool,
         other: (mut telemetry, block_import, grandpa_link, frontier_backend, storage_override),
     } = new_partial(&config, &eth_config, build_import_queue)?;
-
-    consensus_mechanism.spawn_essential_handles(
-        &mut task_manager,
-        client.clone(),
-        custom_service_signal,
-    )?;
 
     let FrontierPartialComponents {
         filter_pool,
@@ -432,15 +530,21 @@ where
             prometheus_registry.clone(),
         ));
 
-        let slot_duration = consensus_mechanism.slot_duration(&client)?;
-        let pending_create_inherent_data_providers =
-            move |_, ()| async move { CM::create_inherent_data_providers(slot_duration) };
+        let slot_duration = sc_consensus_aura::slot_duration(&*client)?;
+        let pending_create_inherent_data_providers = move |_, ()| async move {
+            let current = sp_timestamp::InherentDataProvider::from_system_time();
+            let next_slot = current
+                .timestamp()
+                .as_millis()
+                .saturating_add(slot_duration.as_millis());
+            let timestamp = sp_timestamp::InherentDataProvider::new(next_slot.into());
+            let slot = sp_consensus_aura::inherents::InherentDataProvider::from_timestamp_and_slot_duration(
+				*timestamp,
+				slot_duration,
+			);
+            Ok((slot, timestamp))
+        };
 
-        let rpc_methods = consensus_mechanism.rpc_methods(
-            client.clone(),
-            keystore_container.keystore(),
-            select_chain.clone(),
-        )?;
         Box::new(move |subscription_task_executor| {
             let eth_deps = crate::rpc::EthDeps {
                 client: client.clone(),
@@ -479,8 +583,6 @@ where
                 deps,
                 subscription_task_executor,
                 pubsub_notification_sinks.clone(),
-                CM::frontier_consensus_data_provider(client.clone())?,
-                rpc_methods.as_slice(),
             )
             .map_err(Into::into)
         })
@@ -522,7 +624,7 @@ where
                 sealing,
                 client,
                 transaction_pool,
-                select_chain.clone(),
+                select_chain,
                 block_import,
                 &task_manager,
                 prometheus_registry.as_ref(),
@@ -542,13 +644,18 @@ where
             telemetry.as_ref().map(|x| x.handle()),
         );
 
-        let slot_duration = consensus_mechanism.slot_duration(&client)?;
-        let create_inherent_data_providers =
-            move |_, ()| async move { CM::create_inherent_data_providers(slot_duration) };
+        let slot_duration = sc_consensus_aura::slot_duration(&*client)?;
+        let create_inherent_data_providers = move |_, ()| async move {
+            let timestamp = sp_timestamp::InherentDataProvider::from_system_time();
+            let slot = sp_consensus_aura::inherents::InherentDataProvider::from_timestamp_and_slot_duration(
+				*timestamp,
+				slot_duration,
+			);
+            Ok((slot, timestamp))
+        };
 
-        consensus_mechanism.start_authoring(
-            &mut task_manager,
-            crate::consensus::StartAuthoringParams {
+        let aura = sc_consensus_aura::start_aura::<AuraPair, _, _, _, _, _, _, _, _, _, _>(
+            sc_consensus_aura::StartAuraParams {
                 slot_duration,
                 client,
                 select_chain,
@@ -560,11 +667,17 @@ where
                 force_authoring,
                 backoff_authoring_blocks,
                 keystore: keystore_container.keystore(),
-                block_proposal_slot_portion: SlotProportion::new(2f32 / 3f32),
+                block_proposal_slot_portion: sc_consensus_aura::SlotProportion::new(2f32 / 3f32),
                 max_block_proposal_slot_portion: None,
                 telemetry: telemetry.as_ref().map(|x| x.handle()),
+                compatibility_mode: sc_consensus_aura::CompatibilityMode::None,
             },
         )?;
+        // the AURA authoring task is considered essential, i.e. if it
+        // fails we take down the service with it.
+        task_manager
+            .spawn_essential_handle()
+            .spawn_blocking("aura", Some("block-authoring"), aura);
     }
 
     if enable_grandpa {
@@ -618,35 +731,22 @@ where
     Ok(task_manager)
 }
 
-pub async fn build_full<CM: ConsensusMechanism>(
+pub async fn build_full(
     config: Configuration,
     eth_config: EthConfiguration,
     sealing: Option<Sealing>,
-    custom_service_signal: Option<Arc<AtomicBool>>,
 ) -> Result<TaskManager, ServiceError> {
     match config.network.network_backend {
         sc_network::config::NetworkBackendType::Libp2p => {
-            new_full::<sc_network::NetworkWorker<_, _>, CM>(
-                config,
-                eth_config,
-                sealing,
-                custom_service_signal,
-            )
-            .await
+            new_full::<sc_network::NetworkWorker<_, _>>(config, eth_config, sealing).await
         }
         sc_network::config::NetworkBackendType::Litep2p => {
-            new_full::<sc_network::Litep2pNetworkBackend, CM>(
-                config,
-                eth_config,
-                sealing,
-                custom_service_signal,
-            )
-            .await
+            new_full::<sc_network::Litep2pNetworkBackend>(config, eth_config, sealing).await
         }
     }
 }
 
-pub fn new_chain_ops<CM: ConsensusMechanism>(
+pub fn new_chain_ops(
     config: &mut Configuration,
     eth_config: &EthConfiguration,
 ) -> Result<
@@ -660,7 +760,6 @@ pub fn new_chain_ops<CM: ConsensusMechanism>(
     ServiceError,
 > {
     config.keystore = sc_service::config::KeystoreConfig::InMemory;
-    let mut consensus_mechanism = CM::new();
     let PartialComponents {
         client,
         backend,
@@ -668,7 +767,7 @@ pub fn new_chain_ops<CM: ConsensusMechanism>(
         task_manager,
         other,
         ..
-    } = new_partial(config, eth_config, consensus_mechanism.build_biq()?)?;
+    } = new_partial(config, eth_config, build_aura_grandpa_import_queue)?;
     Ok((client, backend, import_queue, task_manager, other.3))
 }
 
@@ -697,7 +796,7 @@ fn run_manual_seal_authorship(
     thread_local!(static TIMESTAMP: RefCell<u64> = const { RefCell::new(0) });
 
     /// Provide a mock duration starting at 0 in millisecond for timestamp inherent.
-    /// Each call will increment timestamp by slot_duration making the consensus logic think time has passed.
+    /// Each call will increment timestamp by slot_duration making Aura think time has passed.
     struct MockTimestampInherentDataProvider;
 
     #[async_trait::async_trait]
