+++ pallets/subtensor/Cargo.toml
@@ -157,6 +157,4 @@ try-runtime = [
 	"pallet-crowdloan/try-runtime",
 ]
 pow-faucet = []
-fast-blocks = [
-	"subtensor-runtime-common/fast-blocks"
-]
+fast-blocks = []
