+++ runtime/Cargo.toml
@@ -43,7 +43,6 @@ frame-metadata-hash-extension.workspace = true
 sp-api.workspace = true
 sp-block-builder.workspace = true
 sp-consensus-aura.workspace = true
-sp-consensus-babe.workspace = true
 sp-core.workspace = true
 sp-storage.workspace = true
 sp-genesis-builder.workspace = true
@@ -54,7 +53,6 @@ sp-session.workspace = true
 sp-std.workspace = true
 sp-transaction-pool.workspace = true
 sp-version.workspace = true
-sp-authority-discovery.workspace = true
 subtensor-runtime-common.workspace = true
 subtensor-precompiles.workspace = true
 
@@ -93,27 +91,6 @@ pallet-registry.workspace = true
 
 # Metadata commitment pallet
 pallet-commitments.workspace = true
-
-# NPoS
-frame-election-provider-support = { workspace = true }
-pallet-authority-discovery = { workspace = true }
-pallet-babe = { workspace = true }
-pallet-bags-list = { workspace = true }
-pallet-staking = { workspace = true }
-pallet-staking-runtime-api = { workspace = true }
-pallet-session = { workspace = true }
-pallet-fast-unstake = { workspace = true }
-pallet-nomination-pools = { workspace = true }
-pallet-nomination-pools-runtime-api = { workspace = true }
-pallet-offences = { workspace = true }
-pallet-authorship = { workspace = true }
-pallet-election-provider-multi-phase = { workspace = true }
-pallet-staking-reward-fn = { workspace = true }
-pallet-staking-reward-curve = { workspace = true }
-sp-staking = { workspace = true }
-sp-npos-elections = { workspace = true }
-sp-consensus-slots = { workspace = true }
-sp-consensus-grandpa = { workspace = true }
 
 # Frontier
 fp-evm.workspace = true
@@ -174,6 +151,11 @@ std = [
 	"frame-system/std",
 	"frame-try-runtime/std",
 	"pallet-subtensor/std",
+	"pallet-subtensor-swap/std",
+	"pallet-subtensor-swap-runtime-api/std",
+	"substrate-fixed/std",
+	"subtensor-swap-interface/std",
+	"pallet-aura/std",
 	"pallet-balances/std",
 	"pallet-grandpa/std",
 	"pallet-insecure-randomness-collective-flip/std",
@@ -191,6 +173,7 @@ std = [
 	"precompile-utils/std",
 	"sp-api/std",
 	"sp-block-builder/std",
+	"sp-consensus-aura/std",
 	"sp-core/std",
 	"sp-inherents/std",
 	"sp-offchain/std",
@@ -214,24 +197,6 @@ std = [
 	"subtensor-precompiles/std",
 	"subtensor-runtime-common/std",
 	"pallet-crowdloan/std",
-	"pallet-babe/std",
-	"pallet-session/std",
-	"pallet-staking/std",
-	"pallet-election-provider-multi-phase/std",
-	"frame-election-provider-support/std",
-	"pallet-authority-discovery/std",
-	"pallet-authorship/std",
-	"pallet-bags-list/std",
-	"pallet-fast-unstake/std",
-	"pallet-nomination-pools-runtime-api/std",
-	"pallet-nomination-pools/std",
-	"pallet-offences/std",
-	"pallet-staking-reward-fn/std",
-	"pallet-staking-runtime-api/std",
-	"sp-authority-discovery/std",
-	"sp-consensus-babe/std",
-	"sp-npos-elections/std",
-	"sp-staking/std",
 	# Frontier
 	"fp-evm/std",
 	"fp-rpc/std",
@@ -254,14 +219,6 @@ std = [
 	"rand_chacha/std",
 	"sha2/std",
 	"w3f-bls/std",
-	"pallet-aura/std",
-	"sp-consensus-aura/std",
-	"sp-consensus-grandpa/std",
-	"sp-consensus-slots/std",
-	"substrate-fixed/std",
-	"pallet-subtensor-swap/std",
-	"pallet-subtensor-swap-runtime-api/std",
-	"subtensor-swap-interface/std",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking/runtime-benchmarks",
@@ -271,10 +228,12 @@ runtime-benchmarks = [
 	"pallet-balances/runtime-benchmarks",
 	"pallet-grandpa/runtime-benchmarks",
 	"pallet-timestamp/runtime-benchmarks",
+	"pallet-transaction-payment/runtime-benchmarks",
 	"pallet-utility/runtime-benchmarks",
 	"sp-runtime/runtime-benchmarks",
 	"pallet-safe-mode/runtime-benchmarks",
 	"pallet-subtensor/runtime-benchmarks",
+	"pallet-subtensor-swap/runtime-benchmarks",
 	"pallet-collective/runtime-benchmarks",
 	"pallet-membership/runtime-benchmarks",
 	"pallet-proxy/runtime-benchmarks",
@@ -286,29 +245,19 @@ runtime-benchmarks = [
 	"pallet-scheduler/runtime-benchmarks",
 	"pallet-sudo/runtime-benchmarks",
 	"pallet-crowdloan/runtime-benchmarks",
-	"pallet-babe/runtime-benchmarks",
-	"pallet-staking/runtime-benchmarks",
-	"pallet-election-provider-multi-phase/runtime-benchmarks",
-	"frame-election-provider-support/runtime-benchmarks",
-	"pallet-bags-list/runtime-benchmarks",
-	"pallet-fast-unstake/runtime-benchmarks",
-	"pallet-nomination-pools/runtime-benchmarks",
-	"pallet-offences/runtime-benchmarks",
-	"sp-staking/runtime-benchmarks",
 
 	# EVM + Frontier
 	"pallet-ethereum/runtime-benchmarks",
 	"pallet-evm/runtime-benchmarks",
 	"pallet-hotfix-sufficients/runtime-benchmarks",
 	"pallet-drand/runtime-benchmarks",
-	"pallet-transaction-payment/runtime-benchmarks",
-	"pallet-subtensor-swap/runtime-benchmarks",
 ]
 try-runtime = [
 	"frame-try-runtime/try-runtime",
 	"frame-executive/try-runtime",
 	"frame-system/try-runtime",
 	"frame-support/try-runtime",
+	"pallet-aura/try-runtime",
 	"pallet-sudo/try-runtime",
 	"pallet-balances/try-runtime",
 	"pallet-grandpa/try-runtime",
@@ -330,17 +279,6 @@ try-runtime = [
 	"pallet-commitments/try-runtime",
 	"pallet-registry/try-runtime",
 	"pallet-crowdloan/try-runtime",
-	"pallet-babe/try-runtime",
-	"pallet-session/try-runtime",
-	"pallet-staking/try-runtime",
-	"pallet-election-provider-multi-phase/try-runtime",
-	"frame-election-provider-support/try-runtime",
-	"pallet-authority-discovery/try-runtime",
-	"pallet-authorship/try-runtime",
-	"pallet-bags-list/try-runtime",
-	"pallet-fast-unstake/try-runtime",
-	"pallet-nomination-pools/try-runtime",
-	"pallet-offences/try-runtime",
 
 	# EVM + Frontier
 	"fp-self-contained/try-runtime",
@@ -349,6 +287,5 @@ try-runtime = [
 	"pallet-evm/try-runtime",
 	"pallet-evm-chain-id/try-runtime",
 	"pallet-drand/try-runtime",
-	"pallet-aura/try-runtime",
 ]
 metadata-hash = ["substrate-wasm-builder/metadata-hash"]
